netdog: move to utilize dogtag for generate-hostname and update dogtag to handle retry
