+++ sources/netdog/Cargo.toml
@@ -13,11 +13,9 @@ default = []
 wicked = []
 
 [dependencies]
-anyhow = "1.0"
 argh = "0.1"
-dns-lookup = "2"
+dogtag = { version = "0.1", path = "../dogtag" }
 ipnet = { version = "2", features = ["serde"] }
-imdsclient = { path = "../imdsclient", version = "0.1" }
 indexmap = { version = "1", features = ["serde"] }
 envy = "0.4"
 lazy_static = "1"
@@ -30,7 +28,6 @@ serde_json = "1"
 serde_plain = "1"
 snafu = "0.8"
 tokio = { version = "~1.32", default-features = false, features = ["macros", "rt-multi-thread", "time"] }  # LTS
-tokio-retry = "0.3"
 toml = { version = "0.8", features = ["preserve_order"] }
 
 [dev-dependencies]
@@ -38,5 +35,4 @@ tempfile = "3"
 handlebars = "4"
 
 [build-dependencies]
-bottlerocket-variant = { version = "0.1", path = "../bottlerocket-variant" }
 generate-readme = { version = "0.1", path = "../generate-readme" }
